+++ cyberdrop_dl/constants.py
@@ -59,22 +59,26 @@ class CustomHTTPStatus(IntEnum):
     DDOS_GUARD = 429
 
 
-BLOCKED_DOMAINS = (
-    "facebook",
-    "instagram",
-    "fbcdn",
-    "gfycat",
-    "ko-fi.com",
-    "paypal.me",
-    "amazon.com",
-    "throne.com",
-    "youtu.be",
-    "youtube.com",
-    "linktr.ee",
-    "beacons.page",
-    "beacons.ai",
-    "allmylinks.com",
-)
+class BlockedDomains:
+    partial_match = (
+        "facebook",
+        "twitter.com",
+        "instagram",
+        "fbcdn",
+        "gfycat",
+        "ko-fi.com",
+        "paypal.me",
+        "amazon.com",
+        "throne.com",
+        "youtu.be",
+        "youtube.com",
+        "linktr.ee",
+        "beacons.page",
+        "beacons.ai",
+        "allmylinks.com",
+        ".x.com",
+    )
+    exact_match = ("x.com",)
 
 
 DEFAULT_APP_STORAGE = Path("./AppData")