+++ tests/crawlers/test_cases/missav.py
@@ -1,0 +1,43 @@
+DOMAIN = "missav"
+TEST_CASES = [
+    (
+        "https://missav.ws/en/search/coral",
+        [
+            {
+                "url": "re:surrit",
+                "download_folder": r"re:coral \[search\] \(MissAV\)",
+                "album_id": None,
+            },
+        ],
+        3,
+    ),
+    ("https://missav.ws/en/search/MGK", [], 19),
+    (
+        "https://missav.ws/en/gana-3278",
+        [
+            {
+                "url": "https://surrit.com/37f04ee2-0345-40b8-b878-7251db60b236/playlist.m3u8",
+                "filename": "GANA-3278 Futako Tamagawa x Job-Hunting Student x Naturally Fair-Skinned Beauty Suc [1080p].mp4",
+                "original_filename": "GANA-3278 \"Futako Tamagawa x Job-Hunting Student x Naturally Fair-Skinned Beauty: Successfully Picking Up a Pure Girl with a Ladylike Atmosphere!? At First She Was a Little Embarrassed and Blushing, Saying, 'This is My First Time...' But the Moment She Got in the Bed, She Turned On...\": MGS Video (Prestige Group) Adult Video Streaming Site",
+                "referer": "https://missav.ws/en/gana-3278",
+                "download_folder": r"re:Loose Files \(MissAV\)",
+                "album_id": None,
+                "datetime": 1761062400,
+            },
+        ],
+    ),
+    (
+        "https://missav.ws/dm13/en/simm-806",
+        [
+            {
+                "url": "https://surrit.com/5ab04f22-facf-437c-89af-3971da1c0d26/playlist.m3u8",
+                "filename": "SIMM-806 Moeka (18) - A lewd young lady who uses amazing erotic techniques and is ta [720p].mp4",
+                "original_filename": "SIMM-806 Moeka (18) / A lewd young lady who uses amazing erotic techniques and is talented and splayed with an excellent brain J ♪ [1st period] Legs spread open → Enchanted by playing with my favorite clitoris ♪ A high deviation value vulgar blowjob with a perfect score So the neat mouth is full of jichin! A large amount of pregnancy confirmed in a raw vagina that is good at begging! [2nd period] Equipped with black cat lingerie to heat up even more! ? \"I won't stop. You can't stop♪\" Here again, I show off my erotic talent and show off my joints to the provocative woman on top posture. !",
+                "referer": "https://missav.ws/en/simm-806",
+                "download_folder": r"re:Loose Files \(MissAV\)",
+                "album_id": None,
+                "datetime": 1676476800,
+            },
+        ],
+    ),
+]