+++ CHANGELOG.md
@@ -22,6 +22,31 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.5.0] - 2025-10-25
+
+### Added
+
+- Support for static images (RedGifs)
+- Support for JPG7 URLS (JPG5)
+- Support for Transflix
+- Support for individual files within folders (PixelDrain)
+- Support for subfolders in filesystem URLs (PixelDrain)
+- Support collections, tags and search URLs (MissAV)
+- Update supported domains and proxies (PixelDrain)
+
+### Changed
+
+- Improve hash query performance
+- Add rate limit back (Gofile)
+- New blocked domains: `twitter`, `x.com`
+
+### Fixed
+
+- Some `timeout` errors being reported as `Unknown`
+- Do not crash if webhook request failed
+- Date parsing (eHentai)
+- URL parsing (ImageBam)
+
 ## [8.4.0] - 2025-10-18
 
 ### Added